requirements.txt
PyMuPDF
easyocr
numpy
Pillow
